+++ app/i18n/en.json
@@ -1,0 +1,10 @@
+{
+  "common": {
+    "button": "Button"
+  },
+  "card": {
+    "memberCard": {
+      "stack": "Stacks"
+    }
+  }
+}
